Added /claim after opening /vote to context helper

+++ processing/vote.py
@@ -8,7 +8,7 @@ import discord
 
 from cache import messages
 from database import reminders, users
-from resources import exceptions, functions, regex
+from resources import exceptions, functions, regex, strings
 
 
 async def process_message(message: discord.Message, embed_data: Dict, user: Optional[discord.User],
@@ -39,6 +39,7 @@ async def create_reminder(message: discord.Message, embed_data: Dict, user: Opti
         'click here to vote', #English
     ]
     if any(search_string in embed_data['description'].lower() for search_string in search_strings):
+        reminder = None
         if user is None:
             if embed_data['embed_user'] is not None:
                 user = embed_data['embed_user']
@@ -55,22 +56,35 @@ async def create_reminder(message: discord.Message, embed_data: Dict, user: Opti
             except exceptions.FirstTimeUserError:
                 return add_reaction
         if not user_settings.bot_enabled or not user_settings.reminder_vote.enabled: return add_reaction
+        if user_settings.helper_context_enabled:
+            streak_match = re.search(r'\*\*(\d)\*\*/7', embed_data['field1']['value'])
+            if streak_match:
+                await user_settings.update(streak_vote=int(streak_match.group(1)))
         if 'cooldown ready!' in embed_data['title'].lower():
-            try:
-                reminder = await reminders.get_reminder(user.id, 'vote')
+            if reminder is None:
+                try:
+                    reminder = await reminders.get_reminder(user.id, 'vote')
+                except exceptions.NoDataFoundError:
+                    pass
+            if reminder is not None:
                 await reminder.delete()
-                return add_reaction
-            except exceptions.NoDataFoundError:
-                pass
-            return add_reaction
-        user_command = await functions.get_game_command(user_settings, 'vote')
-        timestring_match = re.search(r'\*\*`(.+?)`\*\*', embed_data['title'], re.IGNORECASE)
-        time_left = await functions.calculate_time_left_from_timestring(message, timestring_match.group(1))
-        if time_left < timedelta(0): return add_reaction
-        reminder_message = user_settings.reminder_vote.message.replace('{command}', user_command)
-        reminder: reminders.Reminder = (
-            await reminders.insert_reminder(user.id, 'vote', time_left,
-                                            message.channel.id, reminder_message)
-        )
-        if user_settings.reactions_enabled and reminder.record_exists: add_reaction = True
+        else:
+            user_command = await functions.get_game_command(user_settings, 'vote')
+            timestring_match = re.search(r'\*\*`(.+?)`\*\*', embed_data['title'], re.IGNORECASE)
+            time_left = await functions.calculate_time_left_from_timestring(message, timestring_match.group(1))
+            if time_left < timedelta(0): return add_reaction
+            reminder_message = user_settings.reminder_vote.message.replace('{command}', user_command)
+            reminder: reminders.Reminder = (
+                await reminders.insert_reminder(user.id, 'vote', time_left,
+                                                message.channel.id, reminder_message)
+            )
+            if user_settings.reactions_enabled and reminder.record_exists: add_reaction = True
+        if user_settings.helper_context_enabled:
+            if reminder is None:
+                try:
+                    reminder = await reminders.get_reminder(user.id, 'vote')
+                except exceptions.NoDataFoundError:
+                    pass
+            if reminder is None and user_settings.streak_vote == 7:
+                await message.reply(f"➜ {strings.SLASH_COMMANDS['claim']}")
     return add_reaction
